Query the LiPDverse

Query the LiPDverse

queryLipdverse(
  variable.name = NULL,
  archive.type = NULL,
  paleo.proxy = NULL,
  paleo.units = NULL,
  coord = c(-90, 90, -180, 180),
  age.min = NULL,
  age.max = NULL,
  pub.info = NULL,
  country = NULL,
  continent = NULL,
  ocean = FALSE,
  seasonality = NULL,
  season.not = NULL,
  interp.vars = NULL,
  interp.details = NULL,
  compilation = NULL,
  verbose = FALSE,
  skip.update = FALSE
)

Arguments

variable.name

vector of variable names from c(d18O, d13C, treeRing, etc.)

archive.type

vector of archive types. see unique(queryTable1$archiveType)

paleo.proxy

proxy type c(d18O, mg/ca)

paleo.units

proxy units c(per mil, degC)

coord

lat/lon extent of interest c(latMin, latMax, lonMin, lonMax)

age.min

The youngest portion of the record comes up to at least this age, age BP

age.max

The oldest part of the record extends to at least this age, age BP

pub.info

last name of author, DOI, unique word from title, etc. all lowercase, based on associated publications

country

Coutnry origin of dataset from unique(queryTable1$country), based on lat/lon

continent

Continent origin of dataset from unique(queryTable1$continent), based on lat/lon

ocean

Gather datasets from the marine environment, based on lat/lon

seasonality

list of seasons where items within a list are treated with "AND" logic and separate lists are treated with "OR" logic ie. list(list("July", "August"), list("7,8"), list("summer"))

season.not

seasons not desired with input format identical to seasonality

interp.vars

vector of interpretation variables ie. c("SST", "upwelling"), see possible: unique(queryTable$interp_Vars)

interp.details

vector of interpretation variables ie. c("sea@surface", "elNino"), see possible: unique(queryTable$interp_Details),

compilation

#compilation name as character vector from unique(queryTable$paleoData_mostRecentCompilations), eg. c("temp12k", "wnam")

verbose

offer details of filters and get feedback from user

skip.update

remove update prompt

Value

filtered query table